signlang26thmarch/sign-language/requirements.txt
numpy
pandas
opencv-python
mediapipe
pillow
tensorflow
scikit-learn
pyttsx3
